makes socket generation more robust
we need to have dedicated ports to receive responses on, so that we can
map these responses to channel ids. these ports should be random to mask
users who open a lot of channels to the same destination, otherwise they
would have the same port every time.
to keep track of ports in use, we introduce a list just for that and a
function which deals with finding an available random port.
the function also creates a socket and checks if binding is successful.