puts CBC into own module and creates function to get CBC from keys
I want to maybe test different block cipher operation modes and make
them available. For that I want as little trouble in changing out the
algorithm as possible. That's why EntryPoint imports the cipher from the
Mix, making it the only place where we would have to change something.
If all works out, all we would need to do is change the module we import
default_cipher() from in Mix.py and the rest just works.